Make sortOrder columns smallints

+++ migrations/tables.ts
@@ -5,7 +5,7 @@ export const arcsTable = (t:Knex.CreateTableBuilder) => {
     t.string("name", 255).notNullable();
     t.string("url", 255).nullable().unique();
     t.bigInteger("parentId").unsigned().references("comicArcs.id").onDelete("SET NULL").nullable();
-    t.integer("sortOrder").notNullable().defaultTo(0);
+    t.smallint("sortOrder").notNullable().defaultTo(0);
     t.boolean("enabled").notNullable().defaultTo(false);
     t.string("thumbnailUrl").nullable();
     t.string("bannerUrl").nullable();
@@ -16,7 +16,7 @@ export const charactersTable = (t:Knex.CreateTableBuilder) => {
     t.bigIncrements();
     t.string("name", 255).notNullable().unique();
     t.string("imageUrl").nullable();
-    t.integer("sortOrder").notNullable().defaultTo(0);
+    t.smallint("sortOrder").notNullable().defaultTo(0);
     t.text("bio").nullable();
 }
 
@@ -25,7 +25,7 @@ export const characterAttributesTable = (t:Knex.CreateTableBuilder) => {
     t.bigInteger("characterId").unsigned().notNullable().references("comicCharacters.id").onDelete("CASCADE");
     t.string("name", 255).notNullable();
     t.string("value", 255).notNullable();
-    t.integer("sortOrder").notNullable().defaultTo(0);
+    t.smallint("sortOrder").notNullable().defaultTo(0);
     t.unique(["characterId", "name"]);
 }
 
@@ -34,7 +34,7 @@ export const pagesTable = (t:Knex.CreateTableBuilder) => {
     t.string("name", 255).notNullable();
     t.string("url", 255).nullable().unique();
     t.bigInteger("arcId").notNullable();
-    t.integer("sortOrder").notNullable().defaultTo(0);
+    t.smallint("sortOrder").notNullable().defaultTo(0);
     t.string("imageUrl").nullable();
     t.text("transcript").nullable();
     t.boolean("enabled").notNullable().defaultTo(false);
@@ -47,7 +47,7 @@ export const pageCommentaryTable = (t:Knex.CreateTableBuilder) => {
     t.bigInteger("userId").unsigned().notNullable().references("users.id").onDelete("CASCADE");
     t.string("title").nullable();
     t.text("text").notNullable();
-    t.integer("sortOrder").notNullable().defaultTo(0);
+    t.smallint("sortOrder").notNullable().defaultTo(0);
 }
 
 export const pageCharactersTable = (t:Knex.CreateTableBuilder) => {
